Ignore warning about font shape having incorrect series
It is not uncommon to get a LaTeX warning like the following:

LaTeX Warning: Font shape declaration has incorrect series value `mc'.

The warning appears to be harmless. (I saw that on the internet
so it must be true, right?) It is also not trivial to get rid of
since some combination of packages or distributions can cause it.